app/requirements_backend.txt
Flask
google-cloud-spanner